added gf_version, shakemap source, and process times to info.json
